% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze.ma}
\alias{analyze.ma}
\title{MA: fold change versus average abundance}
\usage{
analyze.ma(dataSet, ref = NULL)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{ref}{A string (default = NULL) specifying the reference condition for comparison.
If NULL, all pairwise comparisons are performed.}
}
\value{
A list comprising data frames for each comparison, with each data frame containing
the means of the two compared conditions for each protein, as well as the average and
difference in means.
}
\description{
Compute the fold change (M) and average abundance (A) values from the data for MA plots.
}
