% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze.t}
\alias{analyze.t}
\title{Student's t-test}
\usage{
analyze.t(
  dataSet,
  ref = NULL,
  adjust.method = "none",
  paired = FALSE,
  pool.sd = FALSE,
  saveRes = TRUE
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{ref}{A character string (default = NULL) specifying
the reference condition for comparison.
If NULL, all pairwise comparisons are performed.}

\item{adjust.method}{A character string (default = "none") specifying
the correction method for p-value adjustment:
\itemize{
\item "BH" or its alias "fdr":
\insertCite{benjamini1995controlling;textual}{msDiaLogue}.
\item "BY": \insertCite{benjamini2001control;textual}{msDiaLogue}.
\item "bonferroni": \insertCite{bonferroni1936teoria;textual}{msDiaLogue}.
\item "hochberg": \insertCite{hochberg1988sharper;textual}{msDiaLogue}.
\item "holm": \insertCite{holm1979simple;textual}{msDiaLogue}.
\item "hommel": \insertCite{hommel1988stagewise;textual}{msDiaLogue}.
\item "none": None
}
See \code{\link[stats]{p.adjust}} for more details.}

\item{paired}{A logical value (default = FALSE) specifying whether
to perform a paired test.}

\item{pool.sd}{A logical value (default = FALSE) specifying whether
to use a pooled standard deviation.}

\item{saveRes}{A logical value (default = TRUE) specifying whether
to save the analysis results to the current working directory.}
}
\value{
A list comprising data frames for each comparison, with each data frame
containing the means of the two compared conditions for each protein,
the difference in means, and the p-values.
}
\description{
Perform Student's t-tests on the data.
}
\references{
\insertAllCited{}
}
