% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterOutIn}
\alias{filterOutIn}
\title{Filtering proteins or contaminates}
\usage{
filterOutIn(dataSet, removeList, listName)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{removeList}{A Boolean specifying whether entries
in the following list should be removed or selected.}

\item{listName}{A list of proteins or contaminates to select or to remove}
}
\value{
The function returns a filtered 2d dataframe.
}
\description{
filterOutIn() applies a series of filtering steps to the data set
}
\details{
If contaminates are removed, a .csv file is created with the removed data.
}
