% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterings.R
\name{filterProtein}
\alias{filterProtein}
\title{Filter proteins by gene, accession or description}
\usage{
filterProtein(
  dataSet,
  proteinInformation = "preprocess_protein_information.csv",
  text = c(),
  by = c("gene", "accession", "description", "name"),
  removeList = TRUE,
  saveRm = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{proteinInformation}{The name of the .csv file containing protein information data
(including the path to the file, if needed). The file should include the following 4
columns: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", and
"PG.ProteinNames". This file is automatically generated by the function
\code{\link[msDiaLogue]{preprocessing}}.}

\item{text}{A character vector of text used as the key for selecting or removing.}

\item{by}{A character string specifying the information to which the \code{text} filter
is applied, with allowable options: "gene", "accession" and "description".}

\item{removeList}{A boolean (default = TRUE) specifying whether the list of proteins
should be removed or selected.
\itemize{
\item TRUE: Remove the list of proteins from the data set.
\item FALSE: Remove all proteins not in the list from the data set.
}}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory. This option only works when \code{removeList = TRUE}.}
}
\value{
A filtered 2d dataframe.
}
\description{
Filter the preprocessed dataset by gene, accession, or description.
}
\details{
The function is an extension of the function \code{\link[msDiaLogue]{preprocessing}}
that allows for filtering proteins based on additional information.
}
