% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.knn}
\alias{impute.knn}
\title{Imputation by the k-nearest neighbors algorithm}
\usage{
impute.knn(
  dataSet,
  reportImputing = FALSE,
  k = 10,
  rowmax = 0.5,
  colmax = 0.8,
  maxp = 1500,
  seed = 362436069
)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}

\item{k}{An integer (default = 10) indicating the number of neighbors to be used in the
imputation.}

\item{rowmax}{A scalar (default = 0.5) specifying the maximum percent missing data
allowed in any row. For any rows with more than \code{rowmax}*100\% missing are imputed
using the overall mean per sample.}

\item{colmax}{A scalar (default = 0.8) specifying the maximum percent missing data
allowed in any column. If any column has more than \code{colmax}*100\% missing data, the
program halts and reports an error.}

\item{maxp}{An integer (default = 1500) indicating the largest block of proteins
imputed using the k-nearest neighbors algorithm. Larger blocks are divided by two-means
clustering (recursively) prior to imputation.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the
random number generator for reproducibility.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply imputation to the dataset by the k-nearest neighbors algorithm
\insertCite{troyanskaya2001missing}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
