% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.mice_norm}
\alias{impute.mice_norm}
\title{Imputation by Bayesian linear regression}
\usage{
impute.mice_norm(dataSet, m = 5, seed = 362436069)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{m}{An integer (default = 5) specifying the number of multiple
imputations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for
the random number generator for reproducibility.}
}
\value{
An imputed data frame.
}
\description{
Apply imputation to the data set by Bayesian linear regression
\insertCite{rubin1987multiple,schafer1997analysis,van2011mice}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
