% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.min_global}
\alias{impute.min_global}
\title{Imputation by the global minimum}
\usage{
impute.min_global(dataSet)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}
}
\value{
An imputed data frame.
}
\description{
Apply imputation to the data set by the minimum measured value from
any protein found within the entire data set.
}
