% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.min_local}
\alias{impute.min_local}
\title{Imputation by the local minimum}
\usage{
impute.min_local(dataSet)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}
}
\value{
An imputed data frame.
}
\description{
Apply imputation to the data set by the minimum measured value for
that protein in that condition.
}
