% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.pca_bayes}
\alias{impute.pca_bayes}
\title{Imputation by Bayesian principal components analysis}
\usage{
impute.pca_bayes(dataSet, nPcs = NULL, maxSteps = 100)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{nPcs}{An integer specifying the number of principal components to
calculate. The default is set to the minimum between the number of samples
and the number of proteins.}

\item{maxSteps}{An integer (default = 100) specifying the maximum number of
estimation steps.}
}
\value{
An imputed data frame.
}
\description{
Apply imputation to the data set by Bayesian principal components analysis
\insertCite{oba2003bayesian}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
