% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.pca_prob}
\alias{impute.pca_prob}
\title{Imputation by probabilistic principal components analysis}
\usage{
impute.pca_prob(
  dataSet,
  reportImputing = FALSE,
  nPcs = NULL,
  maxIterations = 1000,
  seed = 362436069
)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}

\item{nPcs}{An integer specifying the number of principal components to calculate. The
default is set to the minimum between the number of samples and the number of proteins.}

\item{maxIterations}{An integer (default = 1000) specifying the maximum number of
allowed iterations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply imputation to the dataset by probabilistic principal components analysis
\insertCite{stacklies2007pcamethods}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
