% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute}
\alias{impute}
\title{Imputation of raw data signals}
\usage{
impute(
  dataSet,
  imputeType = "LocalMinVal",
  reqPercentPresent = 51,
  reportImputing = FALSE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{imputeType}{A string (default = "LocalMinVal") specifying what imputation
method to use.}

\item{reqPercentPresent}{An int (default = 51) specifying the require percent of values
that must be present in a given protein by condition combination for values to be imputed.}

\item{reportImputing}{A Boolean (default = FALSE) specifying whether to provide
a shadow data frame with imputed data labels. Alters the return structure}
}
\value{
If reportImputing is FALSE, the function returns the imputed 2d dataframe.
If reportImputing is TRUE, the function returns a list of the imputed 2d dataframe and a shadow matrix
showing which proteins by replicate were imputed.
}
\description{
impute() applies an imputation method to the data set.
}
\details{
"LocalMinVal" replaces missing values with the lowest value from the protein by condition combination.
"GlobalMinVal" replaces missing values with the lowest overall value in the data set.
}
