% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{meanVariancePlot}
\alias{meanVariancePlot}
\title{Plotting a graph of mean versus variance}
\usage{
meanVariancePlot(datMV, title = "")
}
\arguments{
\item{datMV}{A data frame containing the data signals.}

\item{title}{A string with the desired title for the mean-variance plot.}
}
\value{
An object of class \code{plot}.
}
\description{
Take a set of protein data organized by column, calculate the mean and variance of each
column, and then plot those statistics.
}
