% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(dataSet, applyto = "sample", normalizeType = "median", plot = TRUE)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{applyto}{A character string (default = "sample") specifying
the target of normalization. Available options are:
\enumerate{
\item "sample" or "row": Row-wise normalization.
\item "protein" or "column": Column-wise normalization.
}}

\item{normalizeType}{A character string (default = "median") specifying
the normalization type to apply:
\itemize{
\item "auto": Auto scaling \insertCite{jackson1991user}{msDiaLogue}.
\item "level": Level scaling.
\item "mean": Mean centering.
\item "median": Median centering.
\item "pareto": Pareto scaling.
\item "quant": Quantile normalization \insertCite{bolstad2003comparison}{msDiaLogue}.
\item "range": Range scaling.
\item "vast": Variable stability (VAST) scaling. \insertCite{keun2003improved}{msDiaLogue}.
\item "none": None.
}}

\item{plot}{A logical value (default = TRUE) specifying whether
to plot the boxplot for before and after normalization.}
}
\value{
A normalized data frame.
}
\description{
Apply a specified type of normalization to a data set.
}
\references{
\insertAllCited{}
}
