% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{The preprocessing() function takes as input:
1. The file name of a Protein Quantitative csv report generated by Spectranaut
2. A Boolean specifying whether or not to filter out results that are NaN
3. An integer specifying how many unique peptides are required to include a protein
The function then executes the following:
\enumerate{
\item reads the file
\item if applicable, applies filters
\item provides summary statistics and histogram of all values reported in the
data set
\item selects columns that provide necessary information for analysis
\item re-formats the data to present individual proteins as columns and groups
replicates under each protein
\item stores the data as a data.frame and prints a few columns to provide a
visual example to the user.
Finally, the function returns the pre-processed data.
}}
\usage{
preprocessing(fileName, filterNaN = TRUE, filterUnique = 2)
}
\description{
The preprocessing() function takes as input:
1. The file name of a Protein Quantitative csv report generated by Spectranaut
2. A Boolean specifying whether or not to filter out results that are NaN
3. An integer specifying how many unique peptides are required to include a protein
The function then executes the following:
\enumerate{
\item reads the file
\item if applicable, applies filters
\item provides summary statistics and histogram of all values reported in the
data set
\item selects columns that provide necessary information for analysis
\item re-formats the data to present individual proteins as columns and groups
replicates under each protein
\item stores the data as a data.frame and prints a few columns to provide a
visual example to the user.
Finally, the function returns the pre-processed data.
}
}
