% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scaffold_ListeriaLysates_Data}
\alias{Scaffold_ListeriaLysates_Data}
\title{Scaffold Listeria lysates protein abundance dataset}
\format{
A data frame with 1737 rows and 20 columns.
}
\source{
\href{https://proteomics.uconn.edu/}{Proteomics & Metabolomics Facility, University of Connecticut}
}
\usage{
Scaffold_ListeriaLysates_Data
}
\description{
A protein abundance dataset exported from Scaffold, generated from
\emph{Listeria} lysate samples under four experimental conditions
(6, ctrl, DM, and SIC). Each condition includes three technical replicates.
The dataset is intended for illustrating the workflows implemented in
\pkg{msDiaLogue}.
}
\keyword{datasets}
