% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortcondition.R
\name{sortcondition}
\alias{sortcondition}
\title{Creating a keyed list of conditions to the list of proteins that are present}
\usage{
sortcondition(dataSet)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}
}
\value{
A list of lists.
}
\description{
Create a keyed dictionary, where every unique experimental condition is the label for
a list of every protein that has a value for that condition.
}
