% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spectronaut_HeLa6Mix_Data}
\alias{Spectronaut_HeLa6Mix_Data}
\title{Spectronaut HeLa 6Mix protein abundance dataset}
\format{
A data frame with 672 rows and 16 columns.
}
\source{
\href{https://proteomics.uconn.edu/}{Proteomics & Metabolomics Facility, University of Connecticut}
}
\usage{
Spectronaut_HeLa6Mix_Data
}
\description{
A protein abundance dataset exported from Spectronaut, generated from
a HeLa protein digest spiked with a six-protein digest mixture at
three concentrations (50, 100, and 200 fmol).
Each concentration includes four technical replicates.
The dataset is intended for illustrating the workflows implemented in
\pkg{msDiaLogue}.
}
\keyword{datasets}
