% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Toy_Scaffold_Data}
\alias{Toy_Scaffold_Data}
\title{Toy Scaffold Protein Abundance Dataset}
\format{
A data frame with 1737 rows and 20 columns.
}
\source{
\href{https://proteomics.uconn.edu/}{Proteomics & Metabolomics Facility, University of Connecticut}
}
\usage{
Toy_Scaffold_Data
}
\description{
A small toy dataset exported from Scaffold. The data structure mimics
the output format of Scaffold protein reports and can be used to
demonstrate the workflow of the \pkg{msDiaLogue} package.
}
\keyword{datasets}
