% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Toy_Spectronaut_Data.R
\docType{data}
\name{Toy_Spectronaut_Data}
\alias{Toy_Spectronaut_Data}
\title{Toy Spectronaut Protein Abundance Dataset}
\format{
A data frame with 672 rows and 16 columns.
}
\source{
\href{https://proteomics.uconn.edu/}{Proteomics & Metabolomics Facility, University of Connecticut}
}
\usage{
Toy_Spectronaut_Data
}
\description{
A small toy dataset exported from Spectronaut. The data structure mimics
the output format of Spectronaut protein reports and can be used to
demonstrate the workflow of the \pkg{msDiaLogue} package.
}
\keyword{datasets}
