% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform}
\alias{transform}
\title{Log based Transformation}
\usage{
transform(dataSet, logFold = 2)
}
\description{
@description
transform() applies a transformation to the data to stabilize
the variance.
}
\details{
@param dataSet A data frame containing the data signals
@param  logFold An integer specifying the base for the log transformation

@details
The function executes the following:
\enumerate{
\item Plots the mean-variance relationship (using meanVariancePlot())
\item Log-transforms the data, using the specified base.
\item Plots the mean-variance relationship again for comparison
}

@returns The function returns the transformed data.
}
