% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{vennMain}
\alias{vennMain}
\title{Operate the Venn diagram functions}
\usage{
vennMain(combos.list, showUniversal = FALSE, fileName = "test.tiff")
}
\arguments{
\item{combos.list}{The output from \code{sortCondition}, lists of condition combinations.}

\item{showUniversal}{A boolean specifying whether to return a list of proteins that are
present in every condition.}

\item{fileName}{A string ending in .tiff, specifying the name for the Venn diagram.}
}
\value{
The unions of conditions with each proteins present.
}
\description{
Create a Venn diagram if useful, and provide a universal protein list.
}
