% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.boxplot}
\alias{visualize.boxplot}
\title{Boxplot}
\usage{
visualize.boxplot(dataSet)
}
\arguments{
\item{dataSet}{The 2d data set of data.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a boxplot for the data.
}
