% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.dist}
\alias{visualize.dist}
\title{Abundance distributions}
\usage{
visualize.dist(dataSet)
}
\arguments{
\item{dataSet}{A data frame containing the data signals, or a list of data
frames.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate distribution plots for protein abundance values.
}
\details{
If \code{dataSet} is a single data frame, the function summarizes
the distribution of proteins' average abundance across conditions
and replicates, including both a kernel density estimate and an empirical
cumulative distribution function (ECDF).

If \code{dataSet} is a list of data frames, the function produces comparative
density plots across data sets (e.g., before vs after imputation), stratified
by "R.Condition".
}
