% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.ind}
\alias{visualize.ind}
\title{Graph of individuals}
\usage{
visualize.ind(dataSet, label = TRUE, ellipse = TRUE, ellipse.level = 0.95)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}}.}

\item{label}{A boolean (default = TRUE) specifying whether the active individuals to be
labeled.}

\item{ellipse}{A boolean (default = TRUE) specifying whether to draw ellipses around
the individuals.}

\item{ellipse.level}{A numeric value (default = 0.95) specifying the size of the
concentration ellipse in normal probability.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a graph of individuals for the data.
}
