% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.loading}
\alias{visualize.loading}
\title{Loading plot / graph of variables}
\usage{
visualize.loading(dataSet, axes = c(1, 2), label = TRUE)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}} or \code{\link[msDiaLogue]{analyze.plsda}}.}

\item{axes}{A numeric vector (default = c(1, 2)) specifying
the axes of interest.}

\item{label}{A logical value (default = TRUE) specifying whether
the active variables to be labeled.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a loadings plot (graph of variables) for the data.
}
