% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.ma}
\alias{visualize.ma}
\title{MA plot: plots fold change versus average abundance}
\usage{
visualize.ma(dataSet, M.thres = 1)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.ma}}.}

\item{M.thres}{The absolute threshold value of M (fold-change) (default = 1) used to
plot the two vertical lines (-M.thres and M.thres) on the MA plot.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate an MA plot for the data.
}
