% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.rank}
\alias{visualize.rank}
\title{Rank abundance distribution plot (Whittaker plot)}
\usage{
visualize.rank(
  dataSet,
  listName = c(),
  regexName = c(),
  by = NULL,
  facet = c("Condition", "Replicate"),
  ht.color = "black",
  ht.shape = 17,
  ht.size = 1.5,
  ht.textcolor = "black",
  ht.textsize = 2
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{listName}{A character vector specifying
proteins for exact matching to highlight.}

\item{regexName}{A character vector specifying
proteins for regular expression pattern matching to highlight.}

\item{by}{A character string (default = "PG.ProteinName" for Spectronaut,
default = "AccessionNumber" for Scaffold) specifying
the information to which \code{listName} and/or \code{regexName} filter
is applied. Allowable options include:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccession",
"PG.ProteinDescriptions", and "PG.ProteinName".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and
"AlternateID".
}}

\item{facet}{A character string (default = c("Replicate", "Condition"))
specifying grouping variables for faceting. Allowed values are:
\itemize{
\item "Condition": Abundance values are averaged across replicates.
\item "Replicate": Abundance values are averaged across conditions.
\item c("Condition", "Replicate"): No averaging is performed.
\item c("Replicate", "Condition"): No averaging is performed.
}}

\item{ht.color}{A character string (default = "black") specifying
the point color of highlighted proteins.}

\item{ht.shape}{A numeric value (default = 17) specifying
the point shape of highlighted points.
See [ggplot2: Elegant Graphics for Data Analysis](https://ggplot2-book.org/scales-other.html#sec-scale-shape)
for more details.}

\item{ht.size}{A numeric value (default = 1.5) specifying
the point size of highlighted proteins.}

\item{ht.textcolor}{A character string (default = "black") specifying
the font color of text labels for highlighted proteins.}

\item{ht.textsize}{A numeric value (default = 2) specifying
the font size of text labels for highlighted proteins.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a rank abundance distribution plot, also known as Whittaker plot,
for the data.
}
