% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.scree}
\alias{visualize.scree}
\title{Scree plot}
\usage{
visualize.scree(
  dataSet,
  type = c("bar", "line"),
  bar.color = "gray",
  bar.fill = "gray",
  line.color = "black",
  label = TRUE,
  ncp = 10
)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}}.}

\item{type}{A string (default = c("bar", "line"))specifying the plot type. Allowed
values are "bar" for a barplot, "line" for a line plot, or c("bar", "line") to use both
types.}

\item{bar.color}{Color of the bar outline in the bar plot. Defaults to "gray".}

\item{bar.fill}{Fill color of the bars in the bar plot. Defaults to "gray".}

\item{line.color}{Color of the line and point in the line plot. Defaults to "black".}

\item{label}{A boolean (default = TRUE) specifying whether labels are added at the top
of bars or points to show the information retained by each dimension.}

\item{ncp}{A numeric value (default = 10) specifying the number of dimensions to be
shown.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a scree plot for the data.
}
