% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.target}
\alias{visualize.target}
\title{Target protein abundance plot}
\usage{
visualize.target(
  dataSet,
  type = "bar",
  facet = TRUE,
  listName = c(),
  regexName = c(),
  by = NULL
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{type}{A character string specifying the plot type.
Allowable options include:
\itemize{
\item "bar": Bar chart.
\item "boxplot": Boxplot.
\item "violin": Violin plot.
}}

\item{facet}{A logical value (default = TRUE) specifying whether
to facet the plot by protein.}

\item{listName}{A character vector specifying
proteins for exact matching to highlight.}

\item{regexName}{A character vector specifying
proteins for regular expression pattern matching to highlight.}

\item{by}{A character string (default = "PG.ProteinName" for Spectronaut,
default = "AccessionNumber" for Scaffold) specifying
the information to which \code{listName} and/or \code{regexName} filter
is applied. Allowable options include:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccession",
"PG.ProteinDescriptions", and "PG.ProteinName".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and
"AlternateID".
}}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate bar, box, or violin plots for selected protein(s) of interest.
}
