% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.test}
\alias{visualize.test}
\title{Histograms of fold changes and p-values from test results}
\usage{
visualize.test(dataSet)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.mod_t}}, \code{\link[msDiaLogue]{analyze.t}}, or
\code{\link[msDiaLogue]{analyze.wilcox}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate histograms of fold changes and p-values for the data.
}
