% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.upset}
\alias{visualize.upset}
\title{UpSet plot}
\usage{
visualize.upset(dataSet)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate an UpSet plot for the data.
}
