% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.var}
\alias{visualize.var}
\title{Graph of variables}
\usage{
visualize.var(dataSet, label = TRUE)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}}.}

\item{label}{A boolean (default = TRUE) specifying whether the active variables to be
labeled.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a graph of variables for the data.
}
