% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.vip}
\alias{visualize.vip}
\title{VIP scores plot}
\usage{
visualize.vip(dataSet, comp = 1, num = 10, thres = 1, rel.widths)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.plsda}}.}

\item{comp}{An integer (default = 1) specifying the PLS-DA component to use when
ranking variables.}

\item{num}{An integer (default = 10) specifying the number of top variables (highest
VIP scores) to display.}

\item{thres}{A scalar (default = 1) specifying the vertical dashed line drawn at this
VIP value.}

\item{rel.widths}{A numerical vector specifying the proportion of relative widths for
the plot panels. Defaults to widths chosen based on the range of the top variables' VIP
scores and the number of group conditions.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a variable importance in projection (VIP) scores plot for the data.
}
