% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.volcano}
\alias{visualize.volcano}
\title{Volcano plot}
\usage{
visualize.volcano(dataSet, P.thres = 0.05, F.thres = 1)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.mod_t}}, \code{\link[msDiaLogue]{analyze.t}},
or \code{\link[msDiaLogue]{analyze.wilcox}}.}

\item{P.thres}{A numeric value (default = 0.05) specifying
the p-value threshold used to draw the horizontal line at -log10(\code{P.thres})
on the volcano plot.}

\item{F.thres}{threshold (default = 1) specifying
the absolute fold change threshold used to draw the vertical lines
at \code{-F.thres} and \code{F.thres} on the volcano plot.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a volcano plot for the data.
}
