% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze.pca}
\alias{analyze.pca}
\title{PCA: principal component analysis}
\usage{
analyze.pca(dataSet, center = TRUE, scale = TRUE)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{center}{A boolean (default = TRUE) indicating whether the variables should be
shifted to be zero centered.}

\item{scale}{A boolean (default = TRUE) indicating whether the variables should be
scaled to have unit variance before the analysis takes place.}
}
\value{
A list containing the following components: \describe{
\item{sdev}{The standard deviations of the principal components.}
\item{loadings}{The matrix of variable loadings.}
\item{scores}{The principal component scores.}
\item{center}{The centering used.}
\item{scale}{The scaling used.}
}
}
\description{
Perform a principal component analysis
\insertCite{pearson1901lines,hotelling1933analysis}{msDiaLogue} on the data.
}
\references{
\insertAllCited{}
}
