% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze.plsda}
\alias{analyze.plsda}
\title{PLS-DA: partial least squares discriminant analysis}
\usage{
analyze.plsda(
  dataSet,
  method = "kernelpls",
  ncomp,
  center = TRUE,
  scale = FALSE
)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{method}{A character string (default = "kernelpls") specifying the multivariate
regression method to be used: \itemize{
\item "kernelpls": Kernel algorithm \insertCite{dayal1997improved}{msDiaLogue}.
\item "widekernelpls": Wide kernel algorithm \insertCite{rannar1994pls}{msDiaLogue}.
\item "simpls": SIMPLS  algorithm \insertCite{dejong1993simpls}{msDiaLogue}.
\item "oscorespls": NIPALS algorithm (classical orthogonal scores algorithm) \insertCite{martens1989multivariate}{msDiaLogue}.
}}

\item{ncomp}{An integer specifying the number of components to include in the model.
Defaults to min(n-1, p).}

\item{center}{A boolean (default = TRUE) indicating whether the variables should be
shifted to be zero centered.}

\item{scale}{A boolean (default = FALSE) indicating whether the variables should be
scaled to have unit variance before the analysis takes place.}
}
\value{
A list containing the following components: \describe{
\item{coefficients}{An array of regression coefficients for \code{ncomp} components.
The dimensions are c(nvar, npred, \code{ncomp}), where nvar is the number of variables
X (proteins) and npred is the number of predicted variables Y (conditions).}
\item{scores}{A matrix of scores.}
\item{vips}{A matrix of variable importance in projection (VIP) scores.}
\item{loadings}{A matrix of loadings.}
\item{loading.weights}{A matrix of loading weights.}
\item{Xvar}{A vector with the amount of X-variance explained by each component.}
\item{Xtotvar}{Total variance in X.}
\item{ncomp}{The number of components.}
\item{method}{The method used to fit the model.}
\item{center}{Indicates whether centering was applied to the model.}
\item{scale}{The scaling used.}
\item{model}{The model frame.}
}
}
\description{
Perform a partial least squares discriminant analysis on the data.
}
\references{
\insertAllCited{}
}
