% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{The analyze() function takes as input:
1.
The function then executes the following:
\enumerate{
\item reads the file
\item if applicable, applies filters
\item 
}

Finally, the function returns a data frame of test results.}
\usage{
analyze(dataSummary)
}
\description{
The analyze() function takes as input:
1.
The function then executes the following:
\enumerate{
\item reads the file
\item if applicable, applies filters
\item 
}

Finally, the function returns a data frame of test results.
}
