% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyzing summarized data}
\usage{
analyze(
  dataSet,
  testType = "t-test",
  ref = NULL,
  adjust.method = "none",
  paired = FALSE,
  pool.sd = FALSE
)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{testType}{A string (default = "t-test") specifying which statistical test to use:
\enumerate{
\item "t-test": Student's t-test.
\item "mod.t-test": Empirical Bayes moderated t-test
\insertCite{smyth2004linear}{msDiaLogue}.
\item "wilcox-test": Wilcoxon test.
\item "MA": Output to plot an MA plot.
}}

\item{ref}{A string (default = NULL) specifying the reference condition for comparison.
If NULL, all pairwise comparisons are performed.}

\item{adjust.method}{A string (default = "none") specifying the correction method for
p-value adjustment when \code{testType = "*-test"}: \itemize{
\item "BH" or its alias "fdr": \insertCite{benjamini1995controlling;textual}{msDiaLogue}.
\item "BY": \insertCite{benjamini2001control;textual}{msDiaLogue}.
\item "bonferroni": \insertCite{bonferroni1936teoria;textual}{msDiaLogue}.
\item "hochberg": \insertCite{hochberg1988sharper;textual}{msDiaLogue}.
\item "holm": \insertCite{holm1979simple;textual}{msDiaLogue}.
\item "hommel": \insertCite{hommel1988stagewise;textual}{msDiaLogue}.
\item "none": None
}
See \code{\link[stats]{p.adjust}} for more details.}

\item{paired}{A boolean (default = FALSE) specifying whether or not to perform a paired
test when \code{testType = "t-test"} or \code{testType = "wilcox-test"}.}

\item{pool.sd}{A boolean (default = FALSE) specifying whether or not to use a pooled
standard deviation when \code{testType = "t-test"}.}
}
\value{
\itemize{
\item "t-test", "mod.t-test", "wilcox-test": A list comprising data frames for each
comparison, with each data frame containing the means of the two compared conditions
for each protein, the difference in means, and the p-values. Additionally, a separate
data frame called "total" summarizes the results of multiple comparisons.
\item "MA": A list comprising data frames for each comparison, with each data frame
containing the means of the two compared conditions for each protein, as well as the
average and difference in means.
}
}
\description{
Apply a statistical test to the data.
}
\references{
\insertAllCited{}
}
