% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyzing summarized data}
\usage{
analyze(dataSet, conditions, testType = "t-test")
}
\arguments{
\item{conditions}{A list of length 2, composed of integers, specifying
which conditions to compare}

\item{testType}{A string (default = "t-test") specifying which
statistical test to use.}

\item{summaryData}{The 2d data set of data}
}
\value{
The function returns a 2d dataframe with differences of means and
p-values for every protein across the two conditions.
}
\description{
analyze() applies a statistical test to the data.
}
