% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze}
\alias{analyze}
\title{Analyzing summarized data}
\usage{
analyze(dataSet, conditions, testType = "t-test")
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{conditions}{A string specifying which two conditions to compare. This argument
only works when there are more than two conditions in \code{dataSet}.}

\item{testType}{A string (default = "t-test") specifying which statistical test to use:
\enumerate{
\item "t-test": unequal variance t-test.
\item "mod.t-test": moderated t-test \insertCite{smyth2004linear}{msDiaLogue}.
\item "MA": output to plot an MA plot.
}}
}
\value{
A 2d dataframe includes the following information: \itemize{
\item "t-test" or "mod.t-test": The differences in means and P-values for each protein
between the two conditions. Note that the differences are calculated by subtracting the
mean of the second condition from the mean of the first condition (Condition 1 -
Condition 2).
\item "MA": Protein-wise averages within each condition.
}
}
\description{
Apply a statistical test to the data.
}
\references{
\insertAllCited{}
}
