% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze}
\alias{analyze}
\title{Analyzing summarized data}
\usage{
analyze(dataSet, conditions, testType = "t-test")
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{conditions}{A string specifying which two conditions to compare. This argument
only works when there are more than two conditions in \code{dataSet}.}

\item{testType}{A string (default = "t-test") specifying which statistical test to use:
\enumerate{
\item "t-test": unequal variance t-test.
\item "mod.t-test": moderated t-test.
\item "volcano": output to plot a volcano plot.
\item "MA": output to plot an MA plot.
}}
}
\value{
A 2d dataframe with differences of means and P-values for every protein across
the two conditions, sorted in ascending order by P-value.
}
\description{
Apply a statistical test to the data.
}
