% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamissing.R
\name{dataMissing}
\alias{dataMissing}
\title{Counting missing data}
\usage{
dataMissing(dataSet)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}
}
\value{
A 2d dataframe with the number and percent of missing data for every protein.
}
\description{
Calculate the count of missing data.
}
