% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamissing.R
\name{dataMissing}
\alias{dataMissing}
\title{Counting missing data}
\usage{
dataMissing(
  dataSet,
  sort_miss = FALSE,
  plot = FALSE,
  show_pct_legend = TRUE,
  show_labels = TRUE,
  show_pct_col = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{sort_miss}{A boolean (default = FALSE) specifying whether to arrange the columns
in order of missingness.}

\item{plot}{A boolean (default = FALSE) specifying whether to plot the missingness.}

\item{show_pct_legend}{A boolean (default = TRUE) specifying whether the percentages of
missing and present values in the entire dataset are shown in the legend of the
visualization when \code{plot = TRUE}.}

\item{show_labels}{A boolean (default = TRUE) specifying whether protein names are
shown in the visualization when \code{plot = TRUE}.}

\item{show_pct_col}{A boolean (default = TRUE) specifying whether the percentages of
missing data in the samples for that protein are shown in the labels of the
visualization when \code{show_labels = TRUE}.}
}
\value{
A 2d dataframe including: \itemize{
\item "count_miss": The count of missing values for each protein.
\item "pct_miss_col": The percentage of missing values for each protein.
\item "pct_miss_tot": The percentage of missing values for each protein relative to
the total missing values in the entire dataset.
}
}
\description{
Calculate and plot the missingness.
}
