% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterNA}
\alias{filterNA}
\title{Filtering NA's post-imputation}
\usage{
filterNA(dataSet, saveRm = TRUE)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory.}
}
\value{
A filtered 2d dataframe.
}
\description{
Remove proteins with NA values.
}
\details{
If proteins that do not meet the imputation requirement are removed, a .csv file is
created with the removed data.
}
