% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterNA}
\alias{filterNA}
\title{Filtering NA's post-imputation}
\usage{
filterNA(dataSet)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}
}
\value{
The function returns a filtered 2d dataframe.
}
\description{
filterNA() removes all proteins from the data
}
\details{
If contaminates are removed, a .csv file is created with the removed data.
}
