% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterOutIn}
\alias{filterOutIn}
\title{Filtering proteins or contaminants}
\usage{
filterOutIn(dataSet, removeList, listName = c(), stringSearch = "")
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{removeList}{A boolean specifying whether entries
in the following list should be removed or selected.
"TRUE" will remove the list of proteins from the data set.
"FALSE will remove all proteins not in the list from the data set.}

\item{listName}{A list of proteins or contaminants to select or to remove}

\item{stringSearch}{A string providing to match against protein names for either inclusion or exclusion}
}
\value{
The function returns a filtered 2d dataframe.
}
\description{
filterOutIn() applies a series of filtering steps to the data set
}
\details{
If proteins are removed by list, a .csv file is created with the removed data.
}
