% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterings.R
\name{filterOutIn}
\alias{filterOutIn}
\title{Filtering proteins or contaminants}
\usage{
filterOutIn(
  dataSet,
  listName = c(),
  regexName = c(),
  by = NULL,
  removeList = TRUE,
  saveRm = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{listName}{A character vector of text strings used as keys for selecting or
removing.}

\item{regexName}{A character vector specifying proteins for regular expression pattern
matching to select or remove.}

\item{by}{A character string (default = "PG.ProteinName" for Spectronaut, default =
"AccessionNumber" for Scaffold) specifying the information to which \code{listName}
and/or \code{regexName} filter is applied. Allowable options include:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccession", "PG.ProteinDescriptions", and
"PG.ProteinName".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".
}}

\item{removeList}{A boolean (default = TRUE) specifying whether the list of proteins
should be removed or selected.
\itemize{
\item TRUE: Remove the list of proteins from the data set.
\item FALSE: Remove all proteins not in the list from the data set.
}}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory. This option only works when \code{removeList = TRUE}.}
}
\value{
A filtered 2d dataframe.
}
\description{
Apply a series of filtering steps to the data set.
}
\details{
If both \code{listName} and \code{regexName} are provided, the proteins selected or
removed will be the union of those specified in \code{listName} and those matching
the regex pattern in \code{regexName}.
}
