% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.knn_seq}
\alias{impute.knn_seq}
\title{Imputation by the k-nearest neighbors algorithm}
\usage{
impute.knn_seq(dataSet, reportImputing = FALSE, k = 10)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}

\item{k}{An integer (default = 10) indicating the number of neighbors to be used in the
imputation.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply imputation to the dataset by the sequential k-nearest neighbors algorithm
\insertCite{kim2004reuse}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
