% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.knn_trunc}
\alias{impute.knn_trunc}
\title{Imputation by the truncated k-nearest neighbors algorithm}
\usage{
impute.knn_trunc(dataSet, k = 10)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{k}{An integer (default = 10) indicating the number of neighbors to be used in the
imputation.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by the truncated k-nearest neighbors algorithm
\insertCite{shah2017distribution}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
