% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.mice_cart}
\alias{impute.mice_cart}
\title{Imputation by classification and regression trees}
\usage{
impute.mice_cart(dataSet, reportImputing = FALSE, m = 5, seed = 362436069)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}

\item{m}{An integer (default = 5) specifying the number of multiple imputations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply imputation to the dataset by classification and regression trees
\insertCite{breiman1984classification,doove2014recursive,van2018flexible}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
