% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.mice_cart}
\alias{impute.mice_cart}
\title{Imputation by classification and regression trees}
\usage{
impute.mice_cart(dataSet, m = 5, seed = 362436069)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{m}{An integer (default = 5) specifying the number of multiple imputations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by classification and regression trees
\insertCite{breiman1984classification,doove2014recursive,van2018flexible}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
