% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.min_global}
\alias{impute.min_global}
\title{Imputation by the global minimum}
\usage{
impute.min_global(dataSet)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by the minimum measured value from any protein found
within the entire dataset.
}
