% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.min_local}
\alias{impute.min_local}
\title{Imputation by the local minimum}
\usage{
impute.min_local(dataSet)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by the minimum measured value for that protein in that
condition.
}
