% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.min_local}
\alias{impute.min_local}
\title{Imputation by the local minimum}
\usage{
impute.min_local(dataSet, reportImputing = FALSE, reqPercentPresent = 0.51)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}

\item{reqPercentPresent}{A scalar (default = 0.51) specifying the required percent of
values that must be present in a given protein by condition combination for values to
be imputed.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply imputation to the dataset by the minimum measured value for that protein in that
condition.
}
