% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.nuc_norm}
\alias{impute.nuc_norm}
\title{Imputation by the nuclear-norm regularization}
\usage{
impute.nuc_norm(
  dataSet,
  rank.max = NULL,
  lambda = NULL,
  thresh = 1e-05,
  maxit = 100,
  final.svd = TRUE,
  seed = 362436069
)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{rank.max}{An integer specifying the restriction on the rank of the solution. The
default is set to one less than the minimum dimension of the dataset.}

\item{lambda}{A scalar specifying the nuclear-norm regularization parameter. If
\code{lambda = 0}, the algorithm convergence is typically slower. The default is set to
the maximum singular value obtained from the singular value decomposition (SVD) of the
dataset.}

\item{thresh}{A scalar (default = 1e-5) specifying the convergence threshold, measured
as the relative change in the Frobenius norm between two successive estimates.}

\item{maxit}{An integer (default = 100) specifying the maximum number of iterations
before the convergence is reached.}

\item{final.svd}{A boolean (default = TRUE) specifying whether to perform a one-step
unregularized iteration at the final iteration, followed by soft-thresholding of the
singular values, resulting in hard zeros.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the
random number generator for reproducibility.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by the nuclear-norm regularization
\insertCite{hastie2015matrix}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
