% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations.R
\name{impute.pca_prob}
\alias{impute.pca_prob}
\title{Imputation by probabilistic principal components analysis}
\usage{
impute.pca_prob(dataSet, nPcs = NULL, maxIterations = 1000, seed = 362436069)
}
\arguments{
\item{dataSet}{The 2d dataset of experimental values.}

\item{nPcs}{An integer specifying the number of principal components to calculate. The
default is set to the minimum between the number of samples and the number of proteins.}

\item{maxIterations}{An integer (default = 1000) specifying the maximum number of
allowed iterations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
An imputed 2d dataframe.
}
\description{
Apply imputation to the dataset by probabilistic principal components analysis
\insertCite{stacklies2007pcamethods}{msDiaLogue}.
}
\references{
\insertAllCited{}
}
