% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute}
\alias{impute}
\title{Imputation of raw data signals}
\usage{
impute(dataSet, imputeType = "MinVal")
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{imputeType}{A Boolean (default = "MinVal") specifying what imputation
method to use.
.}
}
\value{
The function returns an imputed 2d dataframe.
}
\description{
impute() applies an imputation method to the data set.
}
