% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute}
\alias{impute}
\title{Imputation of raw data signals}
\usage{
impute(dataSet, imputeType = "LocalMinVal", reqPercentPresent = 51)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{imputeType}{A Boolean (default = "LocalMinVal") specifying what imputation
method to use.}

\item{reqPercentPresent}{An int (default = 51) specifying the require percent of values
that must be present in a given protein by condition combination for values to be imputed.}
}
\value{
The function returns an imputed 2d dataframe.
}
\description{
impute() applies an imputation method to the data set.
}
\details{
"LocalMinVal" replaces missing values with the lowest value from the protein by condition combination.
"GlobalMinVal" replaces missing values with the lowest overall value in the data set.
}
