% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{Imputation of raw data signals}
\usage{
impute(
  dataSet,
  imputeType = "LocalMinVal",
  reqPercentPresent = 0.51,
  k = 10,
  rowmax = 0.5,
  colmax = 0.8,
  maxp = 1500,
  rng.seed = 362436069,
  reportImputing = FALSE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{imputeType}{A string (default = "LocalMinVal") specifying which imputation
method to use:
\enumerate{
\item "LocalMinVal": replace missing values with the lowest value from the protein by
condition combination.
\item "GlobalMinVal": replace missing values with the lowest value found within the
entire data set.
\item "knn": replace missing values using the k-nearest neighbors algorithm
(Troyanskaya et al., 2001).
\item "seq-knn": replace missing values using the sequential k-nearest neighbors
algorithm (Kim et al., 2004).
\item "trunc-knn": replace missing values using the truncated k-nearest neighbors
algorithm (Shah et al., 2017).
}}

\item{reqPercentPresent}{An numeric value (default = 0.51) specifying the required
percent of values that must be present in a given protein by condition combination for
values to be imputed when \code{imputeType = "LocalMinVal"}.}

\item{k}{An integer (default = 10) indicating the number of neighbors to be used in the
imputation when \code{imputeType} is \code{"knn"}, \code{"seq-knn"}, or
\code{"trunc-knn"}.}

\item{rowmax}{A numeric value (default = 0.5) specifying the maximum percent missing
data allowed in any row when \code{imputeType = "knn"}. For any rows with more than
\code{rowmax}*100\% missing are imputed using the overall mean per sample.}

\item{colmax}{A numeric value (default = 0.8) specifying the maximum percent missing
data allowed in any column when \code{imputeType = "knn"}. If any column has more than
\code{colmax}*100\% missing data, the program halts and reports an error.}

\item{maxp}{An integer (default = 1500) indicating the largest block of proteins
imputed using the k-nearest neighbors algorithm when \code{imputeType = "knn"}. Larger
blocks are divided by two-means clustering (recursively) prior to imputation.}

\item{rng.seed}{An integer (default = 362436069) specifying the seed used for the
random number generator for reproducibility when \code{imputeType = "knn"}.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels, where 1 indicates the corresponding entries
have been imputed, and 0 indicates otherwise. Alters the return structure.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply an imputation method to the data set.
}
\references{
\itemize{
\item Troyanskaya, O., Cantor, M., Sherlock, G., Brown, P., Hastie, T., Tibshirani, R.,
Botstein, D. and Altman, R. B. (2001).
Missing Value Estimation Methods for DNA Microarrays.
\emph{Bioinformatics}, 17(6), 520--525.
\item Kim, KY., Kim, BJ. and Yi, GS. (2004).
Reuse of Imputed Data in Microarray Analysis Increases Imputation Efficiency.
\emph{BMC Bioinformatics}, 5, 160.
\item Shah, J. S., Rai, S. N., DeFilippis, A. P., Hill, B. G., Bhatnagar, A. and
Brock, G. N. (2017).
Distribution Based Nearest Neighbor Imputation for Truncated High Dimensional Data with
Applications to Pre-Clinical and Clinical Metabolomics Studies.
\emph{BMC Bioinformatics}, 18, 114.
}
}
