% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute}
\alias{impute}
\title{Imputation of raw data signals}
\usage{
impute(
  dataSet,
  imputeType = "LocalMinVal",
  reqPercentPresent = 51,
  reportImputing = FALSE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{imputeType}{A string (default = "LocalMinVal") specifying which imputation
method to use:
\enumerate{
\item "LocalMinVal": replace missing values with the lowest value from the protein by
condition combination.
\item "GlobalMinVal": replaces missing values with the lowest value found within the
entire data set.
}}

\item{reqPercentPresent}{An integer (default = 51) specifying the required percentage
of values that must be present in a given protein by condition combination for values
to be imputed.}

\item{reportImputing}{A boolean (default = FALSE) specifying whether to provide a
shadow data frame with imputed data labels. Alters the return structure.}
}
\value{
\itemize{
\item If \code{reportImputing = FALSE}, the function returns the imputed 2d dataframe.
\item If \code{reportImputing = TRUE}, the function returns a list of the imputed 2d
dataframe and a shadow matrix showing which proteins by replicate were imputed.
}
}
\description{
Apply an imputation method to the data set.
}
