% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{meanVariancePlot}
\alias{meanVariancePlot}
\title{Plotting a Mean vs. Variance graph}
\usage{
meanVariancePlot(datMV, title = "")
}
\arguments{
\item{datMV}{A data frame containing the data signals}

\item{title}{A string with the desired title for the mean-variance plot}
}
\value{
The function does not return anything.
}
\description{
meanVariancePlot() takes a set of protein data organized by column, calculates
the mean and variance of each column and that plots those statistics.
}
