% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization of preprocessed data}
\usage{
normalize(dataSet, normalizeType = "Quant")
}
\arguments{
\item{dataSet}{The 2d data set of experimental values}

\item{normalizeType}{A string (default = "Quant") specifying which type of
normalization to apply:
Quantile ("Quant"),
Column-wise Median ("Median"),
Column-wise Mean ("Mean"),
or None ("None").}
}
\value{
The function returns a normalized 2d dataframe.
}
\description{
normalize() applies one specified type of normalization to a data set
}
\details{
Quantile normalization is generally recommended. Mean and Median
normalization are included as popular previous methods. No normalization is not recommended.
}
