% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalization of preprocessed data}
\usage{
normalize(dataSet, normalizeType = "quant")
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{normalizeType}{A string (default = "quant") specifying which type of
normalization to apply:
\enumerate{
\item "quant": Quantile (Smyth, 2004)
\item "median": Protein-wise Median
\item "mean": Protein-wise Mean
\item "none": None
}}
}
\value{
A normalized 2d dataframe.
}
\description{
Apply a specified type of normalization to a data set.
}
\details{
Quantile normalization is generally recommended. Mean and median normalization are
going to be included as popular previous methods. No normalization is not recommended.
Boxplots are also generated for before and after the normalization to give a visual
indicator of the changes.
}
\references{
Smyth, Gordon K. (2004).
Linear Models and Empirical Bayes Methods for Assessing Differential Expression in
Microarray Experiments.
\emph{Statistical Applications in Genetics and Molecular Biology}, 3(1).
}
