% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization of preprocessed data}
\usage{
normalize(dataSet, applyto = "sample", normalizeType = "quant")
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{applyto}{A string (default = "sample") specifying the target of normalization.
Options are "sample" or "row" (across rows), and "protein" or "column" (across columns).}

\item{normalizeType}{A string (default = "quant") specifying which type of
normalization to apply:
\itemize{
\item "auto": Auto scaling.
\item "level": Level scaling.
\item "mean": Mean centering.
\item "median": Median centering.
\item "pareto": Pareto scaling.
\item "quant": Quantile normalization \insertCite{bolstad2003comparison}{msDiaLogue}.
\item "range": Range scaling.
\item "vast": Vast scaling.
\item "none": None.
}}
}
\value{
A normalized 2d dataframe.
}
\description{
Apply a specified type of normalization to a data set.
}
\details{
Quantile normalization is generally recommended. Mean and median normalization are
going to be included as popular previous methods. No normalization is not recommended.
Boxplots are also generated for before and after the normalization to give a visual
indicator of the changes.
}
\references{
\insertAllCited{}
}
