% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{preProcessFiltering}
\alias{preProcessFiltering}
\title{Filtering of raw data signals}
\usage{
preProcessFiltering(
  dataSet,
  filterNaN = TRUE,
  filterUnique = 2,
  replaceBlank = TRUE,
  saveRm = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of experimental values.}

\item{filterNaN}{A boolean (default = TRUE) specifying whether proteins with NaN should
be removed from the data set.}

\item{filterUnique}{An integer (default = 2) specifying whether proteins with less than
the default number of unique peptides should be removed from the data set.}

\item{replaceBlank}{A boolean (default = TRUE) specifying whether proteins without names
should be be named by their accession numbers.}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory.}
}
\value{
A filtered 2d dataframe.
}
\description{
Apply a series of filtering steps to the data set.
}
\details{
All forms of filtering are recommended for most use cases.
}
