% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Loading, filtering and reformatting of MS DIA data}
\usage{
preprocessing(fileName, filterNaN = TRUE, filterUnique = 2, filterBlank = TRUE)
}
\arguments{
\item{fileName}{The name of the file containing MS data (+ the path to the file, if needed)}

\item{filterNaN}{A boolean (default = TRUE) specifying if observations including NaN should be omitted.}

\item{filterUnique}{An integer specifying how many unique peptides are required to include a protein}
}
\value{
The function returns a 2d dataframe.
}
\description{
preprocessing() reads a .csv file, applies filtering conditions, selects columns
necessary for analysis, and returns the reformated data.
}
\details{
The function executes the following:
\enumerate{
\item reads the file
\item if applicable, applies filters
\item provides summary statistics and histogram of all values reported in the
data set
\item selects columns that provide necessary information for analysis
\item re-formats the data to present individual proteins as columns and groups
replicates under each protein
\item stores the data as a data.frame and prints a few columns to provide a
visual example to the user.
}
}
