% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Loading, filtering and reformatting of MS DIA data}
\usage{
preprocessing(
  fileName,
  filterNaN = TRUE,
  filterUnique = 2,
  replaceBlank = TRUE,
  saveRm = TRUE
)
}
\arguments{
\item{fileName}{The name of the file containing MS data (including the path to the file,
if needed).}

\item{filterNaN}{A boolean (default = TRUE) specifying whether observations including
NaN should be omitted.}

\item{filterUnique}{An integer (default = 2) specifying how many number of unique
peptides are required to include a protein.}

\item{replaceBlank}{A boolean (default = TRUE) specifying whether proteins without names
should be be named by their accession numbers.}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory.}
}
\value{
A 2d dataframe.
}
\description{
Read a .csv file, apply filtering conditions, select columns necessary for analysis,
and return the reformatted data.
}
\details{
The function executes the following:
\enumerate{
\item Reads the file.
\item Applies applicable filters, if necessary.
\item Provides summary statistics and a histogram of all values reported in the data set.
\item Selects columns that contain necessary information for the analysis.
\item Re-formats the data to present individual proteins as columns and group replicates
under each protein.
\item Stores the data as a \code{data.frame} and prints a subset of columns to provide
a visual example to the user.
}
}
