% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullProteinPath.R
\name{pullProteinPath}
\alias{pullProteinPath}
\title{Compiling data on a single protein from each step in the process}
\usage{
pullProteinPath(
  listName = NULL,
  regexName = NULL,
  dataSetList,
  proteinInformation = "preprocess_protein_information.csv"
)
}
\arguments{
\item{listName}{A character vector identifying the proteins of interest.}

\item{regexName}{A character vector specifying the proteins for regular expression
pattern matching.}

\item{dataSetList}{A list of data frames, the order dictates the order of presentation.}

\item{proteinInformation}{The name of the .csv file containing protein information data
(including the path to the file, if needed). This file is automatically generated by
the function \code{\link[msDiaLogue]{preprocessing}}.}
}
\value{
A 2d dataframe, with the protein data at each step present in the \code{dataSetList}.
}
\description{
Summarize the steps performed on the data for one protein.
}
