% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullProteinPath.R
\name{pullProteinPath}
\alias{pullProteinPath}
\title{Compiling data on a single protein from each step in the process}
\usage{
pullProteinPath(
  listName = NULL,
  regexName = NULL,
  dataSetList,
  by = "PG.ProteinNames",
  proteinInformation = "preprocess_protein_information.csv"
)
}
\arguments{
\item{listName}{A character vector identifying the proteins of interest.}

\item{regexName}{A character vector specifying the proteins for regular expression
pattern matching.}

\item{dataSetList}{A list of data frames, the order dictates the order of presentation.}

\item{by}{A character string specifying the information to which the \code{listName}
and/or \code{regexName} are applied, with allowable options:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions",
and "PG.ProteinNames".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".
}}

\item{proteinInformation}{The name of the .csv file containing protein information data
(including the path to the file, if needed). The file should include the following
columns:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions",
and "PG.ProteinNames".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".
}
This file is automatically generated by the function
\code{\link[msDiaLogue]{preprocessing}} or
\code{\link[msDiaLogue]{preprocessing_scaffold}}.}
}
\value{
A 2d dataframe, with the protein data at each step present in the \code{dataSetList}.
}
\description{
Summarize the steps performed on the data for one protein.
}
