% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullProteinPath.R
\name{pullProteinPath}
\alias{pullProteinPath}
\title{Compiling data on a single protein from each step in the process}
\usage{
pullProteinPath(proteinName, dataSetList)
}
\arguments{
\item{proteinName}{A String identifying the protein of interest.}

\item{dataSetList}{A list of data frames, the order dictates the order of presentation}
}
\value{
The function returns a 2d dataframe, with the protein data at each step present in the dataSetList
}
\description{
pullProteinPath summarizes the steps that have been performed on the data for one protein.
}
\details{
proteinName must match the labels in the data sets exactly.
}
