% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullProteinPath.R
\name{pullProteinPath}
\alias{pullProteinPath}
\title{Compiling data on a single protein from each step in the process}
\usage{
pullProteinPath(proteinName, dataSetList)
}
\arguments{
\item{proteinName}{A string identifying the protein of interest.}

\item{dataSetList}{A list of data frames, the order dictates the order of presentation.}
}
\value{
A 2d dataframe, with the protein data at each step present in the \code{dataSetList}.
}
\description{
Summarize the steps performed on the data for one protein.
}
\details{
\code{proteinName} must match the labels in the data sets exactly.
}
