% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize protein intensities across conditions}
\usage{
summarize(dataSet, saveSumm = TRUE)
}
\arguments{
\item{dataSet}{A data frame containing the data signals and labels.}

\item{saveSumm}{A boolean (default = TRUE) specifying whether to save the summary
statistics to current working directory.}
}
\value{
A 2d summarized data frame.
}
\description{
Calculate the mean, standard deviation, and replicate count for protein across every
condition.
}
\details{
The column 'Stat' in the generated data.frame includes the following statistics:
\itemize{
\item n: Number.
\item mean: Mean.
\item sd: Standard deviation.
\item median: Median.
\item trimmed: Trimmed mean with a trim of 0.1.
\item mad: Median absolute deviation (from the median).
\item min: Minimum.
\item max: Maximum.
\item range: The difference between the maximum and minimum value.
\item skew: Skewness.
\item kurtosis: Kurtosis.
\item se: Standard error.
}
}
