% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarize}
\alias{summarize}
\title{Summarize protein intensities across replicates}
\usage{
summarize(dataSet, fileName = "")
}
\arguments{
\item{dataSet}{A data frame containing the data signals and labels}

\item{fileName}{A string which can be used to generate a csv file of the summary statistics}
}
\value{
The function returns a 2d summarized data frame.
}
\description{
summarize() calculates the mean, standard deviation and replicate count for
protein across every condition.
}
