% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform}
\alias{transform}
\title{Log-based transformation}
\usage{
transform(dataSet, logFold = 2)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{logFold}{An integer specifying the base for the log transformation.}
}
\value{
The transformed data.
}
\description{
Apply a logarithmic transformation to the data to stabilize the variance.
}
\details{
The function executes the following:
\enumerate{
\item Plots the mean-variance relationship using \code{meanVariancePlot()}.
\item Log-transforms the data, using the specified base.
\item Plots the mean-variance relationship again for comparison.
}
}
