% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transformation}
\usage{
transform(dataSet, method = "log", logFold = 2, root = 2)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{method}{A string (default = "log") specifying the method to be used for the
transformation:
\enumerate{
\item "log": Logarithm transformation.
\item "root": Root transformation.
}}

\item{logFold}{An integer (default = 2) specifying the base for the log transformation
when \code{method = "log"}.}

\item{root}{An integer (default = 2) specifying the degree of the root for the root
transformation when \code{method = "root"}. For example, set it to 2 for square root or
3 for cube root.}
}
\value{
The transformed data.
}
\description{
Apply a transformation to the data to stabilize the variance.
}
\details{
The function executes the following:
\enumerate{
\item Plots the mean-variance relationship.
\item Transforms the data.
\item Plots the mean-variance relationship again for comparison.
}
}
