% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFASTA.R
\name{trimFASTA}
\alias{trimFASTA}
\title{Trimming down a protein FASTA file to certain proteins}
\usage{
trimFASTA(
  FASTAFileName,
  reportFileName,
  outputFileName,
  selectString = "*BOVIN"
)
}
\arguments{
\item{FASTAFileName}{A string indicating the FASTA .txt filename.}

\item{reportFileName}{A string indicating the spectronaut report .csv filename.}

\item{outputFileName}{A string indicating the name for the new .txt FASTA file.}

\item{selectString}{A string containing a regular expression for which to search.}
}
\value{
The function returns a FASTA file with only the specified proteins present.
}
\description{
Trimming down a FASTA file to only contain proteins present in an associated
spectronaut report file.
}
\details{
Depending on the size of the FASTA files, this function can be slow and
take multiple minutes to run.
The FASTA file must be in a .txt format. No other FASTA formats will work.
}
