% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{tTesting}
\alias{tTesting}
\title{Producing results of t-test}
\usage{
tTesting(x)
}
\arguments{
\item{x}{The function takes a parameter x, a 1d dataframe, where the first
half of the data are from the first condition, and the second half are
from the second condition.}
}
\value{
The function returns a list of length 2 with the difference of means and
the t-tested p-values for the protein across two conditions.
}
\description{
This function produces output (difference of means, and P value) for visual
inspection or for external use.
}
