% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{tTesting}
\alias{tTesting}
\title{Producing results of t-test}
\usage{
tTesting(x, lengthA)
}
\arguments{
\item{x}{A 1d dataframe, where data from the first condition are listed first, and data
from the second condition are listed second.}

\item{lengthA}{An integer specifying the number of the data points in \code{x} are from
the first condition.}
}
\value{
A list of length 2 with the difference of means and the P-values from a t-test
for the protein across two conditions.
}
\description{
Produce output (difference of means, and P-value) for visual inspection or for external
use.
}
