% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{tTesting}
\alias{tTesting}
\title{Producing results of t-test}
\usage{
tTesting(x)
}
\arguments{
\item{x}{placeholder}
}
\value{
The function returns a 1d dataframe with differences of means and
p-values across the two conditions.
}
\description{
placeholder
}
