% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.dist}
\alias{visualize.dist}
\title{Average abundance distributions: Density and ECDF}
\usage{
visualize.dist(dataSet)
}
\arguments{
\item{dataSet}{The 2d data set of data.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate distribution plots of proteins' average abundance across conditions
and replicates, including both a kernel density estimate and an empirical
cumulative distribution function (ECDF).
}
