% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.heatmap}
\alias{visualize.heatmap}
\title{Heatmap}
\usage{
visualize.heatmap(
  dataSet,
  pkg = "pheatmap",
  cluster_cols = TRUE,
  cluster_rows = FALSE,
  show_colnames = TRUE,
  show_rownames = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{pkg}{A string (default = "pheatmap") specifying the source package used to plot
the heatmap. Two options: \code{"pheatmap"} and \code{"ggplot2"}.}

\item{cluster_cols}{A boolean (default = TRUE) determining if rows should be clustered
or \code{hclust} object. This argument only works when \code{pkg = "pheatmap"}.}

\item{cluster_rows}{A boolean (default = FALSE) determining if columns should be
clustered or \code{hclust} object. This argument only works when \code{pkg = "pheatmap"}.}

\item{show_colnames}{A boolean (default = TRUE) specifying if column names are be shown.
This argument only works when \code{pkg = "pheatmap"}.}

\item{show_rownames}{A boolean (default = TRUE) specifying if row names are be shown.
This argument only works when \code{pkg = "pheatmap"}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a heatmap for the data.
}
