% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.loading}
\alias{visualize.loading}
\title{Loading plot / graph of variables}
\usage{
visualize.loading(dataSet, label = TRUE)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}} or \code{\link[msDiaLogue]{analyze.plsda}}.}

\item{label}{A boolean (default = TRUE) specifying whether the active variables to be
labeled.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a loadings plot (graph of variables) for the data.
}
