% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.rank}
\alias{visualize.rank}
\title{Rank abundance distribution plot (Whittaker plot)}
\usage{
visualize.rank(
  dataSet,
  listName = NULL,
  regexName = NULL,
  by = NULL,
  facet = c("Condition", "Replicate"),
  color = "red",
  ...
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{listName}{A character vector of proteins to highlight.}

\item{regexName}{A character vector specifying
proteins for regular expression pattern matching to highlight.}

\item{by}{A character string (default = "PG.ProteinName" for Spectronaut,
default = "AccessionNumber" for Scaffold) specifying
the information to which \code{listName} and/or \code{regexName} filter
is applied.
Allowable options include:
\itemize{
\item For Spectronaut: "PG.Genes", "PG.ProteinAccession",
"PG.ProteinDescriptions", and "PG.ProteinName".
\item For Scaffold: "ProteinDescriptions", "AccessionNumber", and
"AlternateID".
}}

\item{facet}{A character string (default = c("Replicate", "Condition"))
specifying grouping variables for faceting. Allowed values are:
\itemize{
\item "Condition": Abundance values are averaged across replicates.
\item "Replicate": Abundance values are averaged across conditions.
\item c("Condition", "Replicate"): No averaging is performed.
\item c("Replicate", "Condition"): No averaging is performed.
}}

\item{color}{A character string (default = red") specifying
the color used to highlight proteins.}

\item{...}{Optional arguments passed to \code{\link[ggrepel]{geom_text_repel}}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a rank abundance distribution plot, also known as Whittaker plot,
for the data.
}
