% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize}
\alias{visualize}
\title{Generating visualizations for MS Data}
\usage{
visualize(
  outputData,
  graphType = "volcano",
  fileName,
  transformType,
  conditionLabels
)
}
\arguments{
\item{outputData}{A 2D data frame that corresponds to the output from the function
\code{analyze()} for the same name.}

\item{graphType}{A string indicating the graph type.
Current options are:
\enumerate{
\item "heatmap"
\item "MA"
\item "normalize"
\item "pca"
\item "t-test"
\item "venn"
\item "volcano"
}}

\item{fileName}{Filename specifying the name for Venn image output, or if NULL, it
returns the grid object itself.}

\item{transformType}{transformType ## TODO}

\item{conditionLabels}{conditionLabels ## TODO}
}
\value{
An object of class \code{ggplot}.
}
\description{
Create specific graphics to illustrate the results of the data analysis function.
}
\details{
The function \code{visualize()} is designed to work directly with output from the
function \code{analyze()}. Please be sure that the arguments \code{graphType} and
\code{testType} match.
}
