% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize}
\alias{visualize}
\title{Generating visualizations for MS Data}
\usage{
visualize(
  outputData,
  graphType = "volcano",
  pkg = "pheatmap",
  cluster_cols = TRUE,
  cluster_rows = FALSE,
  show_colnames = TRUE,
  show_rownames = TRUE,
  compareConditions = c(),
  transformType = NULL,
  conditionLabels = c(),
  P.thres = 0.05,
  logF.thres = 0.6,
  fileName
)
}
\arguments{
\item{outputData}{A 2D data frame that corresponds to the output from the function
\code{analyze()} for the same name.}

\item{graphType}{A string indicating the graph type.
Current options are:
\enumerate{
\item "heatmap"
\item "MA"
\item "normalize"
\item "pca"
\item "t-test"
\item "venn"
\item "volcano"
}}

\item{pkg}{A string specifying the source package used to plot the heatmap.
Two options: \code{"pheatmap"} and \code{"ggplot2"}.
This argument only works when \code{graphType = "heatmap"}.}

\item{cluster_cols}{A boolean (default = TRUE) determining if rows should be clustered
or \code{hclust} object. This argument only works when \code{graphType = "heatmap"} and
\code{pkg = "pheatmap"}.}

\item{cluster_rows}{A boolean (default = FALSE) determining if columns should be
clustered or \code{hclust} object. This argument only works when
\code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{show_colnames}{A boolean (default = TRUE) specifying if column names are be shown.
This argument only works when \code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{show_rownames}{A boolean (default = TRUE) specifying if row names are be shown.
This argument only works when \code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{compareConditions}{A string specifying which two conditions to compare. This
argument only works when there are more than two conditions in \code{outputData}.}

\item{transformType}{A string used to label the title of the transformation type for
the transformed MA plot.}

\item{conditionLabels}{A string used to label the title of the two conditions for the
volcano plot. This argument only works when \code{graphType = "volcano"}.}

\item{P.thres}{THe threshold value of P-value (default = 0.05) used to plot the
horizontal line (-log10(P.thres)) on the volcano plot. This argument only works when
\code{graphType = "volcano"}.}

\item{logF.thres}{The absolute threshold value of log2(fold change) (default = 0.6) used
to plot the two vertical lines (-logF.thres and logF.thres) on the volcano plot. This
argument only works when \code{graphType = "volcano"}.}

\item{fileName}{Filename specifying the name for Venn image output, or if NULL, it
returns the grid object itself.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Create specific graphics to illustrate the results of the data analysis function.
}
\details{
The function \code{visualize()} is designed to work directly with output from the
function \code{analyze()}. Please be sure that the arguments \code{graphType} and
\code{testType} match.
}
