% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize}
\alias{visualize}
\title{Generating visualizations for MS Data}
\usage{
visualize(
  dataSet,
  graphType = "volcano",
  pkg = "pheatmap",
  cluster_cols = TRUE,
  cluster_rows = FALSE,
  show_colnames = TRUE,
  show_rownames = TRUE,
  M.thres = 1,
  transformLabel = "Log2",
  center = TRUE,
  scale = TRUE,
  addlabels = TRUE,
  choice = "variance",
  ncp = 10,
  addEllipses = TRUE,
  ellipse.level = 0.95,
  label = "all",
  show_percentage = TRUE,
  fill_color = c("blue", "yellow", "green", "red"),
  show_universal = FALSE,
  P.thres = 0.05,
  logF.thres = 0.6
)
}
\arguments{
\item{dataSet}{A 2D data frame that corresponds to the output from the function
\code{analyze()} for the same name.}

\item{graphType}{A string indicating the graph type.
Current options are:
\enumerate{
\item "heatmap"
\item "MA"
\item "normalize"
\item "PCA_biplot"
\item "PCA_ind"
\item "PCA_scree"
\item "PCA_var"
\item "t-test"
\item "Upset"
\item "Venn"
\item "volcano"
}}

\item{pkg}{A string specifying the source package used to plot the heatmap.
Two options: \code{"pheatmap"} and \code{"ggplot2"}.
This argument only works when \code{graphType = "heatmap"}.}

\item{cluster_cols}{A boolean (default = TRUE) determining if rows should be clustered
or \code{hclust} object. This argument only works when \code{graphType = "heatmap"} and
\code{pkg = "pheatmap"}.}

\item{cluster_rows}{A boolean (default = FALSE) determining if columns should be
clustered or \code{hclust} object. This argument only works when
\code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{show_colnames}{A boolean (default = TRUE) specifying if column names are be shown.
This argument only works when \code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{show_rownames}{A boolean (default = TRUE) specifying if row names are be shown.
This argument only works when \code{graphType = "heatmap"} and \code{pkg = "pheatmap"}.}

\item{M.thres}{The absolute threshold value of M (log fold-change) (default = 1) used
to plot the two vertical lines (-M.thres and M.thres) on the MA plot when
\code{graphType = "MA"}.}

\item{transformLabel}{A string (default = "Log2") used to label the title and axes of
the transformation type for the transformed MA plot when \code{graphType = "MA"}.}

\item{center}{A boolean (default = TRUE) indicating whether the variables should be
shifted to be zero centered when \code{graphType = "PCA_scree"},
\code{graphType = "PCA_ind"}, \code{graphType = "PCA_var"}, or
\code{graphType = "PCA_biplot"}.}

\item{scale}{A boolean (default = TRUE) indicating whether the variables should be
scaled to have unit variance before the analysis takes place when
\code{graphType = "PCA_scree"}, \code{graphType = "PCA_ind"},
\code{graphType = "PCA_var"}, or \code{graphType = "PCA_biplot"}.}

\item{addlabels}{A boolean (default = TRUE) specifying whether the elements are labeled.
\itemize{
\item For \code{graphType = "PCA_scree"}, it specifies whether labels are added at the
top of bars or points to show the information retained by each dimension.
\item For \code{graphType = "PCA_ind"}, it specifies whether the active individuals to
be labeled.
\item For \code{graphType = "PCA_var"}, it specifies whether the active variables to be
labeled.
}}

\item{choice}{A text (default = "variance") specifying the PCA data to be plotted the
scree plot when \code{graphType = "PCA_scree"}. Allowed values are "variance" or
"eigenvalue".}

\item{ncp}{A numeric value (default = 10) specifying the number of dimensions to be
shown when \code{graphType = "PCA_scree"}.}

\item{addEllipses}{A boolean (default = TRUE) specifying whether to draw ellipses
around the individuals when \code{graphType = "PCA_ind"} or
\code{graphType = "PCA_biplot"}.}

\item{ellipse.level}{A numeric value (default = 0.95) specifying the size of the
concentration ellipse in normal probability when \code{graphType = "PCA_ind"} or
\code{graphType = "PCA_biplot"}.}

\item{label}{A text (default = "all") specifying the elements to be labelled when
\code{graphType = "PCA_biplot"}. Allowed values:
\itemize{
\item "all": Label both active individuals and active variables.
\item "ind": Label only active individuals.
\item "var": Label only active variables.
\item "none": No labels.
}}

\item{show_percentage}{A boolean (default = TRUE) specifying whether to show the
percentage for each set when \code{graphType = "Venn"}.}

\item{fill_color}{A text (default = c("blue", "yellow", "green", "red")) specifying the
colors to fill in circles when \code{graphType = "Venn"}.}

\item{show_universal}{A boolean (default = FALSE) specifying whether to return a
data.frame with logical columns representing sets when \code{graphType = "Venn"}.}

\item{P.thres}{THe threshold value of P-value (default = 0.05) used to plot the
horizontal line (-log10(P.thres)) on the volcano plot when \code{graphType = "volcano"}.}

\item{logF.thres}{The absolute threshold value of log2(fold change) (default = 0.6)
used to plot the two vertical lines (-logF.thres and logF.thres) on the volcano plot
when \code{graphType = "volcano"}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Create specific graphics to illustrate the results of the data analysis function.
}
\details{
The function \code{visualize()} is designed to work directly with output from the
function \code{analyze()}. Please be sure that the arguments \code{graphType} and
\code{testType} match.
}
