% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize}
\alias{visualize}
\title{Generating visualizations for MS Data}
\usage{
visualize(outputData, graphType = "volcano", fileName, transformType)
}
\arguments{
\item{outputData}{A 2d data frame that matches the output from the analyze() function for the same name}

\item{graphType}{A string indicating the graph type}
}
\value{
The function does not return anything.
}
\description{
visualize() produces specified graphics for the results of the data analysis function
}
\details{
visualize() is designed to work directly with output from analyze.
Be sure that the graphType and testType params match.
}
