% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.score}
\alias{visualize.score}
\title{Score plot / graph of individuals}
\usage{
visualize.score(dataSet, ellipse = TRUE, ellipse.level = 0.95, label = TRUE)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.pca}} or \code{\link[msDiaLogue]{analyze.plsda}}.}

\item{ellipse}{A boolean (default = TRUE) specifying whether to draw ellipses around
the individuals.}

\item{ellipse.level}{A numeric value (default = 0.95) specifying the size of the
concentration ellipse in normal probability.}

\item{label}{A boolean (default = TRUE) specifying whether the active individuals to be
labeled.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a scores plot (graph of individuals) for the data.
}
