% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.venn}
\alias{visualize.venn}
\title{Venn diagram}
\usage{
visualize.venn(
  dataSet,
  show_percentage = TRUE,
  fill_color = c("blue", "yellow", "green", "red"),
  saveRes = TRUE
)
}
\arguments{
\item{dataSet}{The 2d data set of data.}

\item{show_percentage}{A boolean (default = TRUE) specifying whether to show the
percentage for each set.}

\item{fill_color}{A text (default = c("blue", "yellow", "green", "red")) specifying the
colors to fill in circles.}

\item{saveRes}{A boolean (default = TRUE) specifying whether to save the data, with
logical columns representing sets, to current working directory.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a Venn diagram for the data.
}
