% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visualize.volcano}
\alias{visualize.volcano}
\title{Volcano plot}
\usage{
visualize.volcano(dataSet, P.thres = 0.05, F.thres = 1)
}
\arguments{
\item{dataSet}{The data set corresponds to the output from the function
\code{\link[msDiaLogue]{analyze.mod_t}}, \code{\link[msDiaLogue]{analyze.t}}, or
\code{\link[msDiaLogue]{analyze.wilcox}}.}

\item{P.thres}{THe threshold value of p-value (default = 0.05) used to plot the
horizontal line (-log10(P.thres)) on the volcano plot.}

\item{F.thres}{The absolute threshold value of fold change (default = 1) used to plot
the two vertical lines (-F.thres and F.thres) on the volcano plot.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Generate a volcano plot for the data.
}
