% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{volcanoTest}
\alias{volcanoTest}
\title{Producing results for a volcano plot}
\usage{
volcanoTest(x, lengthA)
}
\arguments{
\item{x}{A 1d dataframe, where data from the first condition are listed first, and data
from the second condition are listed second.}

\item{lengthA}{An integer specifying the number of the data points in \code{x} are from
the first condition.}
}
\value{
A list of length 2 with the difference of means and the P-values from a t-test
for the protein across two conditions.
}
\description{
Produce output (fold change, and P-value) for a volcano plot.
}
\details{
If the means between the two conditions are the same, t.test throws an error, which is
caught and printed to the user to make sure the program keeps running.

WARNING: Currently, the function does not apply any further transformation. The data
must undergo a log2 transformation in a previous step before being used here.
}
