% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{volcanoTest}
\alias{volcanoTest}
\title{Producing results for volcano plot}
\usage{
volcanoTest(x)
}
\arguments{
\item{x}{placeholder}
}
\value{
The function returns a 1d dataframe with differences of log2 means and
t-tested p-values for every protein across the two conditions.
}
\description{
placeholder
}
